device_agent/requirements.txt
# AetherAI Device Agent — Python Dependencies
# Install on your PC/laptop
websockets>=12.0
pyautogui>=0.9.54
Pillow>=10.0.0
python-dotenv>=1.0.1
